+++ docs/source/ttnn/ttnn/ttnn/erfinv_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.erfinv_bw:
+
+ttnn.erfinv_bw
+###############
+
+.. autofunction:: ttnn.erfinv_bw
