+++ tests/ttnn/unit_tests/operations/backward/test_backward_erfinv.py
@@ -4,8 +4,8 @@
 
 import torch
 import pytest
-import tt_lib
-from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import data_gen_with_range, compare_pcc
+import ttnn
+from tests.ttnn.unit_tests.operations.backward.utility_funcs import data_gen_with_range, compare_pcc
 
 
 @pytest.mark.parametrize(
@@ -21,7 +21,7 @@ def test_bw_erfinv(input_shapes, device):
     grad_data, grad_tensor = data_gen_with_range(input_shapes, -20, 20, device, True)
     in_data.retain_grad()
 
-    tt_output_tensor_on_device = tt_lib.tensor.erfinv_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.erfinv_bw(grad_tensor, input_tensor)
 
     pyt_y = torch.erfinv(in_data)
