+++ tt_eager/tt_dnn/op_library/backward/backward_ops.cpp
@@ -384,46 +384,6 @@ std::vector<Tensor> hardtanh_bw(
     return operation::decorate_as_composite(__func__, _hardtanh_bw)(grad, input, min, max, output_mem_config);
 }
 
-// erfinv
-// self: 0.5 * sqrt(M_PI) * exp(self.erfinv().pow(2)) * grad
-// for input -1 and 1: grad.sign() * inf, for input > 1 or < -1 : nan
-std::vector<Tensor> _erfinv_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    std::vector<Tensor> grad_tensor;
-    Tensor result = ttnn::multiply(
-        ttnn::multiply(ttnn::sqrt(full_like(input, M_PI, output_mem_config), output_mem_config),
-            ttnn::multiply(ttnn::exp(ttnn::square(ttnn::erfinv(input, output_mem_config), output_mem_config), false, output_mem_config),
-                grad,
-                std::nullopt,
-                output_mem_config),
-            std::nullopt,
-            output_mem_config),
-        0.5,
-        std::nullopt,
-        output_mem_config);
-    Tensor neg_one = full_like(input, -1.0, output_mem_config);
-    Tensor pos_one = full_like(input, 1.0, output_mem_config);
-    Tensor t_inf = ttnn::multiply(ttnn::sign(grad, output_mem_config), std::numeric_limits<float>::infinity(), std::nullopt, output_mem_config);
-    result = where(
-        ttnn::logical_or(
-            ttnn::lt(input, neg_one, std::nullopt, output_mem_config),
-            ttnn::gt(input, pos_one, std::nullopt, output_mem_config),
-            std::nullopt,
-            output_mem_config),
-        std::nanf(" "),
-        result,
-        output_mem_config);
-    result = where(
-        ttnn::eq(input, neg_one, std::nullopt, output_mem_config),
-        t_inf,
-        where(ttnn::eq(input, pos_one, std::nullopt, output_mem_config), t_inf, result, output_mem_config),
-        output_mem_config);
-    grad_tensor.emplace_back(result);
-    return grad_tensor;
-}
-std::vector<Tensor> erfinv_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    return operation::decorate_as_composite(__func__, _erfinv_bw)(grad, input, output_mem_config);
-}
-
 std::vector<Tensor> _erf_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
     std::vector<Tensor> grad_tensor;
     Tensor result = ttnn::multiply(
