+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -83,6 +83,7 @@ enum class UnaryBackwardOpType {
     EXPM1_BW,
     RECIPROCAL_BW,
     DIGAMMA_BW,
+    ERFINV_BW,
 };
 
 struct UnaryBackwardFunction{
